Larger messages seem to reduce range dramatically, presumably because a larger message increases the chance for flipped bits or errors. To combat this, I'm introducing multiple message types, sending only the data necessary for the specific message type. Previously we were sending one giant message with a superset of data.
To do this, we need a header, which is 8 bits:

MSB  00000000  LSB
     RRRAAAMM

MM - message type
     00 - contains sensor data message
     01 - pairing, contains keys message
     10 - reserved
     11 - reserved

AAA - sensor address (if applicable to message type)

RRR - reserved (could be message-specific info)

Messages will consist of a 1-byte header, and a 16 byte payload.
There is currently a 'sensordata' payload, and a 'pairingdata'
payload.